backend/requirements.txt
connexion[swagger-ui]==2.3.0
Flask==1.1.1
flask-cors==3.0.8
Flask-SQLAlchemy==2.4.0
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.17.0
psycopg2==2.8.3
pyOpenSSL==19.0.0
